+++ CityBuilder/Game/City.h
@@ -35,7 +35,7 @@ public:
 	inline int Get_Day() const { return (m_DailyTickCounter % 360) % 30 + 1; }
 	inline int Get_Time() const { return m_DailyTickCounter; }
 	inline int Get_GameTableSize() const { return m_GameTable->Get_TableSize(); }
-	inline int Get_CitizenSize() { return m_Citizens.size(); }
+	inline int Get_CitizenSize() { return (int)m_Citizens.size(); }
 	inline std::unordered_set<Citizen*>& Get_Citizens() { return m_Citizens; }
 	inline int Get_NumberOfResidences() const { return m_NumberOfResidences; }
 	inline float Get_CombinedHappiness() { return Get_CitizenSize() == 0 ? 0 : m_combinedHappiness * 100; }
@@ -61,7 +61,7 @@ protected:
 	void GenerateLakes(int iterations, double initialRatio);
 	float CalculateMonthlyTax();
 	void  CollectAnnualCosts();
-	void  UpdateMoney(float amount);
+	void  UpdateMoney(int amount);
 	void SimulatePopulationAging();
 	void SimulateForestAging();
 	void CalculateHappiness();
@@ -73,7 +73,7 @@ private:
 	std::unordered_set<Citizen*> m_Citizens;
 	std::vector<std::pair<int, int>> BresenhamAlgorithm(int x0, int y0, int x1, int y1);
 
-	float m_Money = 0;
+	int m_Money = 0;
 	int m_DailyTickCounter = 0;
 
 	float m_combinedHappiness = 0;
